End of slides update
Now the work is done in MapByRoadComponent and the main class.

The changes on main seem to be good and sturdy, but the invoked later MainWindow seems to be very buggy yet and so it won't run.

For the MapByRoadComponent class it seems to me that it is well made, but I cannot tell for sure since as of now the project is not runnable.