Tables update
Now eventstable is good, vehiclestable seems good, junctionstable and roadstable are still worthless with the previous implementation, also i "fixed" the run method by adding an if that checks if out is null, if it is it does the "GUI mode", we probably should ask for guidance on this fix. The rest of the tables I'll finish in another occasion.