Lists in Dialogs Update
Now the roads and the vehicles appear in the dialogs.

This makes the dialogues' size be too small so I added calls to pack().

Also we must be doing something wrong, running the program consumes memory BIG time. Also the new events created through the GUI don't seem to execute.